+++ src/annotations/core/AnnotationItemFactory.h
@@ -30,6 +30,7 @@
 #include "src/annotations/items/AnnotationRect.h"
 #include "src/annotations/items/AnnotationEllipse.h"
 #include "src/annotations/items/AnnotationNumber.h"
+#include "src/annotations/items/AnnotationNumberPointer.h"
 #include "src/annotations/items/AnnotationText.h"
 #include "src/annotations/items/AnnotationBlur.h"
 #include "src/annotations/items/AnnotationImage.h"