+++ src/widgets/ToolPicker.cpp
@@ -59,29 +59,7 @@ void ToolPicker::initGui()
 	auto button = createButton(action);
 	mLayout->addWidget(button, 0, 0);
 
-	action = createAction(tr("Pen"), IconLoader::load(QStringLiteral("pen.svg")), Qt::Key_P, ToolTypes::Pen);
-	button = createButton(action);
-	mLayout->addWidget(button, 0, 1);
-
-	action = createAction(tr("Number"), IconLoader::load(QStringLiteral("number.svg")), Qt::Key_N, ToolTypes::Number);
-	button = createButton(action);
-	mLayout->addWidget(button, 2, 0);
-
 	auto menu = new QMenu();
-	action = createAction(tr("Marker Rectangle"), IconLoader::load(QStringLiteral("markerRect.svg")), Qt::Key_J, ToolTypes::MarkerRect);
-	menu->addAction(action);
-	action = createAction(tr("Marker Ellipse"), IconLoader::load(QStringLiteral("markerEllipse.svg")), Qt::Key_K, ToolTypes::MarkerEllipse);
-	menu->addAction(action);
-	action = createAction(tr("Marker Pen"), IconLoader::load(QStringLiteral("markerPen.svg")), Qt::Key_M, ToolTypes::MarkerPen);
-	menu->addAction(action);
-	button = createButton(menu);
-	mLayout->addWidget(button, 2, 1);
-
-	action = createAction(tr("Text"), IconLoader::load(QStringLiteral("text.svg")), Qt::Key_T, ToolTypes::Text);
-	button = createButton(action);
-	mLayout->addWidget(button, 3, 0);
-
-	menu = new QMenu();
 	action = createAction(tr("Arrow"), IconLoader::load(QStringLiteral("arrow.svg")), Qt::Key_A, ToolTypes::Arrow);
 	menu->addAction(action);
 	action = createAction(tr("Double Arrow"), IconLoader::load(QStringLiteral("doubleArrow.svg")), Qt::Key_D, ToolTypes::DoubleArrow);
@@ -89,11 +67,37 @@ void ToolPicker::initGui()
 	action = createAction(tr("Line"), IconLoader::load(QStringLiteral("line.svg")), Qt::Key_L, ToolTypes::Line);
 	menu->addAction(action);
 	button = createButton(menu);
-	mLayout->addWidget(button, 3, 1);
+	mLayout->addWidget(button, 0, 1);
+
+	action = createAction(tr("Pen"), IconLoader::load(QStringLiteral("pen.svg")), Qt::Key_P, ToolTypes::Pen);
+	button = createButton(action);
+	mLayout->addWidget(button, 1, 0);
+
+	menu = new QMenu();
+	action = createAction(tr("Marker Rectangle"), IconLoader::load(QStringLiteral("markerRect.svg")), Qt::Key_J, ToolTypes::MarkerRect);
+	menu->addAction(action);
+	action = createAction(tr("Marker Ellipse"), IconLoader::load(QStringLiteral("markerEllipse.svg")), Qt::Key_K, ToolTypes::MarkerEllipse);
+	menu->addAction(action);
+	action = createAction(tr("Marker Pen"), IconLoader::load(QStringLiteral("markerPen.svg")), Qt::Key_M, ToolTypes::MarkerPen);
+	menu->addAction(action);
+	button = createButton(menu);
+	mLayout->addWidget(button, 1, 1);
+
+	action = createAction(tr("Text"), IconLoader::load(QStringLiteral("text.svg")), Qt::Key_T, ToolTypes::Text);
+	button = createButton(action);
+	mLayout->addWidget(button, 2, 0);
+
+	menu = new QMenu();
+	action = createAction(tr("Number"), IconLoader::load(QStringLiteral("number.svg")), Qt::Key_N, ToolTypes::Number);
+	menu->addAction(action);
+	action = createAction(tr("Number Pointer"), IconLoader::load(QStringLiteral("numberPointer.svg")), Qt::Key_O, ToolTypes::NumberPointer);
+	menu->addAction(action);
+	button = createButton(menu);
+	mLayout->addWidget(button, 2, 1);
 
 	action = createAction(tr("Blur"), IconLoader::load(QStringLiteral("blur.svg")), Qt::Key_B, ToolTypes::Blur);
 	button = createButton(action);
-	mLayout->addWidget(button, 4, 0);
+	mLayout->addWidget(button, 3, 0);
 
 	menu = new QMenu();
 	action = createAction(tr("Rectangle"), IconLoader::load(QStringLiteral("rect.svg")), Qt::Key_R, ToolTypes::Rect);
@@ -101,11 +105,11 @@ void ToolPicker::initGui()
 	action = createAction(tr("Ellipse"), IconLoader::load(QStringLiteral("ellipse.svg")), Qt::Key_E, ToolTypes::Ellipse);
 	menu->addAction(action);
 	button = createButton(menu);
-	mLayout->addWidget(button, 4, 1);
+	mLayout->addWidget(button, 3, 1);
 
 	action = createAction(tr("Sticker"), IconLoader::load(QStringLiteral("sticker.svg")), Qt::Key_I, ToolTypes::Sticker);
 	button = createButton(action);
-	mLayout->addWidget(button, 5, 0);
+	mLayout->addWidget(button, 4, 0);
 
 	setLayout(mLayout);
 	setFixedSize(sizeHint());